refactor: moves cesta logic to cestaProvider
